Issue #13 - build new folder structure
