+++ package.json
@@ -11,6 +11,8 @@
     "@testing-library/user-event": "^7.1.2",
     "dexie": "^2.0.4",
     "dompurify": "^2.0.7",
+    "node-sass": "^4.13.0",
+    "normalize.css": "^8.0.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
