+++ src/App.scss
@@ -1,0 +1,73 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+html,
+body {
+  width: 100vw;
+  height: 100vh;
+  margin: 0px;
+  padding: 0px;
+}
+
+ul {
+  list-style: None;
+  padding: 0px;
+  margin: 0px;
+}
+
+#root {
+  width: 100vw;
+  height: 100vh;
+}
+
+.App {
+  width: 100vw;
+  height: 100vh;
+  text-align: center;
+  background-color: #f8f8f8;
+}
+
+.dataset-page {
+  display: grid;
+  grid-template-columns: 50% 50%;
+  grid-template-rows: 200px 1fr;
+  grid-template-areas: 'Header Header' 'Columns Joins';
+  width: 100%;
+  height: 100%;
+}
+
+.dataset-header {
+  grid-area: Header;
+}
+
+.dataset-columns {
+  grid-area: Columns;
+}
+
+.dataset-joins {
+  grid-area: Joins;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+
+  ul {
+    overflow-y: auto;
+    height: 500px;
+  }
+}
+
+.tag-selector {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+}
+.selected {
+  font-weight: bold;
+}
