+++ src/components/CategorySelector/CategorySelector.scss
@@ -1,0 +1,9 @@
+.category-selector,
+.tag-selector {
+  box-sizing: border-box;
+  padding: 10px;
+}
+
+.category-list li {
+  cursor: pointer;
+}
