+++ src/components/Dataset/Dataset.scss
@@ -1,0 +1,18 @@
+.dataset {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid;
+  width: 100%;
+  margin-bottom: 20px;
+  -moz-box-shadow: 2px 2px 23px 0px rgba(0, 0, 0, 0.75);
+  webkit-box-shadow: 2px 2px 23px 0px rgba(0, 0, 0, 0.75);
+  box-shadow: 2px 2px 23px 0px rgba(0, 0, 0, 0.75);
+  background-color: white;
+  box-sizing: border-box;
+  padding: 10px 20px;
+  text-align: left;
+}
+
+.dataset-tags {
+  text-align: right;
+}
