+++ src/components/TagSelector/TagSelector.scss
@@ -1,0 +1,7 @@
+.tag-list {
+  overflow-y: auto;
+  list-style: 'none';
+  li {
+    cursor: pointer;
+  }
+}
