+++ src/layout/HomePage/HomePage.scss
@@ -1,0 +1,55 @@
+.home-page {
+  display: grid;
+  grid-template-columns: 300px 1fr;
+  grid-template-columns: 300px 1fr;
+  grid-template-rows: 50% 50%;
+  grid-template-areas: 'Tags Datasets' 'Categories Datasets';
+  width: 100%;
+  height: 100%;
+}
+
+.tags {
+  width: 100%;
+  height: 100%;
+  background-color: white;
+  grid-area: Tags;
+}
+
+.search {
+  width: 100%;
+  box-sizing: border-box;
+  padding: 0px 0px 20px 0px;
+
+  input {
+    width: 80%;
+    box-sizing: border-box;
+    padding: 10px;
+  }
+}
+
+.pagination {
+  width: 100%;
+}
+
+.categories {
+  text-align: left;
+  grid-area: Categories;
+  overflow-y: auto;
+  background-color: white;
+}
+
+.datasets {
+  grid-area: Datasets;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  padding: 20px;
+}
+
+.dataset-list {
+  overflow-y: auto;
+  flex: 1;
+  box-sizing: border-box;
+  padding: 20px;
+}
